requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
yfinance==0.2.54
pandas
numpy 
scikit-learn
ta==0.11.0
vaderSentiment==3.3.2
scipy
groq==0.20.0
python-dotenv==1.0.1
langchain
langchain-community
duckduckgo-search
langchain-core
streamlit
firebase-admin
matplotlib